extract token balance logic into module
