+++ deployables/app/app/balances/index.ts
@@ -1,0 +1,2 @@
+export { getTokenBalances } from './getTokenBalances'
+export type { BalanceResult, TokenBalance } from './types.server'
