+++ deployables/app/app/balances/types.server.ts
@@ -3,3 +3,5 @@ import type Moralis from 'moralis'
 export type BalanceResult = Awaited<
   ReturnType<(typeof Moralis.EvmApi.wallets)['getWalletTokenBalancesPrice']>
 >['result']
+
+export type TokenBalance = BalanceResult[number]
