+++ deployables/app/tsconfig.json
@@ -7,7 +7,8 @@
     "paths": {
       "@/test-utils": ["./test-utils/index.ts"],
       "@/components": ["./app/components/index.ts"],
-      "@/utils": ["./app/utils/index.ts"]
+      "@/utils": ["./app/utils/index.ts"],
+      "@/balances": ["./app/balances/index.ts"]
     }
   },
   "include": [
